distributed_systems/video_processor/requirements.txt
annotated-types==0.7.0
anyio==4.11.0
boto3==1.40.55
botocore==1.40.55
certifi==2025.10.5
cffi==2.0.0
cryptography==46.0.3
deprecation==2.1.0
google-genai==0.3.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jmespath==1.0.1
multidict==6.7.0
packaging==25.0
postgrest==2.22.0
propcache==0.4.1
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
realtime==2.22.0
s3transfer==0.14.0
six==1.17.0
sniffio==1.3.1
storage3==2.22.0
StrEnum==0.4.15
supabase==2.22.0
supabase-auth==2.22.0
supabase-functions==2.22.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.22.0
